fix(ci): unbreak the gate this branch inherited from main — fmt, and the 1500-line ratchet
Three checks were red on this PR and none of them belong to it. `git diff
<main> HEAD` over every failing file comes back empty: the copies here are
byte-identical to main, whose own CI has been red on its last four pushes.
The originally-reported failure —
`verify::tests::both_judge_prompts_mark_the_diff_as_worker_authored_data` —
was already fixed by #1240 and arrived with the merge; what is left is fmt
and the file-size ratchet.

`cargo fmt --all`. Two re-wraps in `pipeline.rs` (a six-argument call and a
struct pattern) and one module reorder in `lib.rs`. `git diff -w` shows no
change inside any line; nothing semantic moves.

`stella-tui/src/render/tests.rs` had reached 1703 lines and is not
grandfathered, so it was a hard block rather than a ceiling bump — and the
guard is explicit that the fix for a new offender is a split, not a baseline
entry. The accessibility block (#936, #935) comes out into
`render/tests/accessibility.rs`, following the `mod thinking;` precedent
already in that file, which was split off for this same reason. The move is
mechanical: the two helpers it carries (`type_into`, `draw_terminal`) have no
caller outside it, and the six tests are one topic — every one of them either
reads the backend cursor or sets `ui.screen_reader`. `tests.rs` lands at 1464.

The five grandfathered ceilings then need `make file-size-update`, run after
`cargo fmt` rather than before: formatting `pipeline.rs` added nine lines, so
the two failures are coupled and the baseline has to be regenerated against
the formatted tree. The regenerated file also *tightens*
`stella-cli/src/agent_tests.rs` from 1814 to 1743, which is the ratchet doing
its job — a raise-only patch would have left that stale exemption in place.
Entry count is unchanged at 34: no new file is grandfathered.

The growth in `driver.rs` (+136), `driver/tests.rs` (+117) and `pipeline.rs`
(+87) is feature code from #1240 and #1214. The guard's stricter advice is to
move new code into its own module, but restructuring another PR's
just-merged work from inside a Context PR is a worse trade than the visible
baseline diff the guard was designed to produce.
